src/auth/requirements.txt
black==22.10.0
flask-migrate==4.0.0
flask-script==2.0.6
pip-chill==1.0.1
psycopg2-binary==2.9.5
pyjwt==2.6.0
python-decouple==3.6
# alembic==1.8.1 # Installed as dependency for flask-migrate
# click==8.1.3 # Installed as dependency for flask, black
# flask==2.2.2 # Installed as dependency for flask-sqlalchemy, flask-script, flask-migrate
# flask-sqlalchemy==3.0.2 # Installed as dependency for flask-migrate
# greenlet==2.0.1 # Installed as dependency for sqlalchemy
# itsdangerous==2.1.2 # Installed as dependency for flask
# jinja2==3.1.2 # Installed as dependency for flask
# mako==1.2.4 # Installed as dependency for alembic
# markupsafe==2.1.1 # Installed as dependency for jinja2, mako, werkzeug
# mypy-extensions==0.4.3 # Installed as dependency for black
# pathspec==0.10.2 # Installed as dependency for black
# platformdirs==2.5.4 # Installed as dependency for black
# sqlalchemy==1.4.44 # Installed as dependency for flask-sqlalchemy, alembic
# tomli==2.0.1 # Installed as dependency for black
# werkzeug==2.2.2 # Installed as dependency for flask
